+++ packages/react/src/components/ErrorBoundary.tsx
@@ -1,7 +1,7 @@
 ﻿import React, { ReactNode, Suspense } from 'react';
 import { ComponentRendering, LayoutServicePageState } from '@sitecore-content-sdk/core/layout';
-import { withSitecoreContext } from '../enhancers/withSitecoreContext';
-import { SitecoreContextValue } from './SitecoreContext';
+import { withSitecore } from '../enhancers/withSitecore';
+import { SitecoreProviderPageContext } from './SitecoreProvider';
 
 type ErrorComponentProps = {
   [prop: string]: unknown;
@@ -9,7 +9,7 @@ type ErrorComponentProps = {
 
 export type ErrorBoundaryProps = {
   children: ReactNode;
-  sitecoreContext: SitecoreContextValue;
+  pageContext: SitecoreProviderPageContext;
   type: string;
   isDynamic?: boolean;
   errorComponent?: React.ComponentClass<ErrorComponentProps> | React.FC<ErrorComponentProps>;
@@ -48,8 +48,8 @@ class ErrorBoundary extends React.Component<ErrorBoundaryProps> {
   showErrorDetails(): boolean {
     return (
       this.isInDevMode() ||
-      this.props.sitecoreContext?.pageState === LayoutServicePageState.Edit ||
-      this.props.sitecoreContext?.pageState === LayoutServicePageState.Preview
+      this.props.pageContext?.pageState === LayoutServicePageState.Edit ||
+      this.props.pageContext?.pageState === LayoutServicePageState.Preview
     );
   }
 
@@ -94,4 +94,4 @@ class ErrorBoundary extends React.Component<ErrorBoundaryProps> {
   }
 }
 
-export default withSitecoreContext()(ErrorBoundary);
+export default withSitecore()(ErrorBoundary);